Refactor: SkillKeywordManager ìinfrastructure 패키지로 원복
